Rendering works sometimes as aspected
The current step is to render the model described but with the current
strategy the alignment depends on the arrangement list of the
connection points - which is wrong. In the root folder there are two
images (working.png and nonsense.png) which shows the rendering output
of the same object. The first is for „a.b is connected to c.d“, the
second for flipped sides of the same statement („c.d is connected to
a.b“).